fix type signature

+++ runtime/web_worker.rs
@@ -271,7 +271,7 @@ pub struct WebWorkerOptions {
   pub blob_store: BlobStore,
   pub broadcast_channel: InMemoryBroadcastChannel,
   pub shared_array_buffer_store: Option<SharedArrayBufferStore>,
-  pub cpu_count: u32,
+  pub cpu_count: usize,
 }
 
 impl WebWorker {
